refactor: move TokenType to own file
